+++ mixtera_integrationtests/checkpointing/test_server_checkpointing.py
@@ -4,7 +4,7 @@ from pathlib import Path
 from time import sleep
 from unittest import TestCase
 
-from integrationtests.utils import TestMetadataParser, setup_test_dataset
+from mixtera_integrationtests.utils import TestMetadataParser, setup_test_dataset
 from loguru import logger
 
 from mixtera.core.client.mixtera_client import MixteraClient, QueryExecutionArgs, ResultStreamingArgs
