+++ mixtera_integrationtests/mixtera_torch_dataset/test_torch_dataset.py
@@ -4,7 +4,7 @@ import tempfile
 from copy import deepcopy
 from pathlib import Path
 
-from integrationtests.utils import REPRODUCIBILITY_ITERATIONS, TestMetadataParser, setup_test_dataset
+from mixtera_integrationtests.utils import REPRODUCIBILITY_ITERATIONS, TestMetadataParser, setup_test_dataset
 from loguru import logger
 
 import torch
